requirements.txt
aiohttp>=3.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiosqlite>=0.19.0
openai>=1.0.0
jieba>=0.42.1
nltk>=3.8.1
loguru>=0.7.0
apscheduler>=3.10.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0
yfinance>=0.2.36
pandas>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx[socks]
asyncpg>=0.29.0
redis>=5.0.0
qdrant-client>=1.9.0
# ── Phase 1: Information Density ──
trafilatura>=1.8.0
readability-lxml>=0.8.0
instructor>=1.4.0
# ── Phase 2: Local ML Migration ──
fastembed>=0.4.0
datasketch>=1.6.0
rapidfuzz>=3.0.0
gliner>=0.2.0
# ── Phase 3: Deep Integration ──
langfuse>=2.0.0
akshare>=1.14.0
transformers>=4.40.0
# ── Phase 4: Advanced Capabilities ──
hdbscan>=0.8.33
sentence-transformers>=3.0.0
